backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==1.16.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pymongo==3.12.0
python-dotenv==1.0.1
Werkzeug==3.1.3
